+++ src/Filament/Architect/CardBlock.php
@@ -2,8 +2,10 @@
 
 namespace Codedor\FilamentArchitect\Filament\Architect;
 
+use Codedor\FilamentArchitect\ArchitectFormats;
 use Codedor\FilamentArchitect\Filament\Components\ButtonComponent;
 use Codedor\MediaLibrary\Filament\AttachmentInput;
+use Codedor\MediaLibrary\Models\Attachment;
 use Filament\Forms\Components\Repeater;
 use Filament\Forms\Components\TextInput;
 use FilamentTiptapEditor\TiptapEditor;
@@ -15,7 +17,8 @@ class CardBlock extends BaseBlock
         return [
             Repeater::make('cards')
                 ->schema([
-                    AttachmentInput::make('image'),
+                    AttachmentInput::make('image')
+                        ->allowedFormats(ArchitectFormats::get()),
                     TextInput::make('title'),
                     TiptapEditor::make('description'),
                     ButtonComponent::make('button'),
@@ -23,5 +26,14 @@ class CardBlock extends BaseBlock
                 ->grid()
                 ->minItems(1),
         ];
+    }
+
+    public function getData(): array
+    {
+        if ($this->data['data']['image']) {
+            $this->data['data']['image'] = Attachment::find($this->data['data']['image']);
+        }
+
+        return $this->data;
     }
 }