+++ CHANGELOG.md
@@ -5,6 +5,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.3.0 - 2023-09-21
+
+### What's changed
+
+#### Bumped
+
+- Media to v0.3.0
+
+## v0.2.5 - 2023-09-20
+
+### What's changed
+
+#### Added
+
+- OEmbed field in the VideoText block
+- Collapsed tabs by default
+
+## v0.2.1 - 2023-09-20
+
+## What's Changed
+
+- Fix MediaBlock getData
+
+**Full Changelog**: https://github.com/codedor/filament-architect/compare/v0.2.0...v0.2.1
+
+## v0.2.0 - 2023-09-05
+
+## What's Changed
+
+- Updated for the v0.2.0 version of media
+
+**Full Changelog**: https://github.com/codedor/filament-architect/compare/v0.1.1...v0.2.0
+
 ## v0.1.1 - 2023-08-24
 
 **Full Changelog**: https://github.com/codedor/filament-architect/compare/v0.1.0...v0.1.1